# docs/api/openapi.yaml
openapi: 3.0.3

info:
  title: VitalCV API
  description: |
    VitalCV is a decentralized verifiable credentials platform built on W3C standards.

    ## Authentication

    This API uses DID (Decentralized Identifier) authentication with Ed25519 signatures.

    ### Authentication Flow

    1. **Request Challenge**: POST `/api/auth/did/challenge` with your DID
    2. **Sign Challenge**: Sign the returned message with your private key
    3. **Verify Signature**: POST `/api/auth/did/verify` with signature
    4. **Use Token**: Include JWT token in `Authorization: Bearer <token>` header

    ## Rate Limiting

    All endpoints are rate limited. Limits are returned in response headers:

    - `X-RateLimit-Limit`: Maximum requests allowed in window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: UTC timestamp when the rate limit resets

    ## Standards Compliance

    - W3C Verifiable Credentials Data Model 1.1
    - W3C Decentralized Identifiers (DIDs) v1.0
    - Ed25519 Signature 2020
    - BBS+ Signatures (selective disclosure)
    - Zero-Knowledge Proofs (zk-SNARKs)

  version: 1.0.0
  contact:
    name: VitalCV Support
    email: support@vitalcv.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.vitalcv.app
    description: Production server

tags:
  - name: Authentication
    description: DID authentication endpoints
  - name: User
    description: User profile management
  - name: Credentials
    description: Verifiable credentials management (coming soon)
  - name: Verification
    description: Credential verification (coming soon)

paths:
  /api/auth/did/challenge:
    post:
      tags:
        - Authentication
      summary: Request authentication challenge
      description: |
        Generate a cryptographic challenge for DID authentication.
        The challenge must be signed with the private key corresponding to the DID.
      operationId: requestChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - did
              properties:
                did:
                  type: string
                  description: Decentralized Identifier (DID)
                  example: "did:ethr:0x1234567890abcdef1234567890abcdef12345678"
            example:
              did: "did:ethr:0x1234567890abcdef1234567890abcdef12345678"
      responses:
        '200':
          description: Challenge generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    type: string
                    description: Hex-encoded challenge
                  message:
                    type: string
                    description: Formatted message to sign
                  expiresAt:
                    type: string
                    format: date-time
                    description: Challenge expiration time (5 minutes)
              example:
                challenge: "a1b2c3d4e5f6..."
                message: "VitalCV Authentication Request\n\nPlease sign this message..."
                expiresAt: "2024-10-08T16:05:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/auth/did/verify:
    post:
      tags:
        - Authentication
      summary: Verify signed challenge
      description: |
        Verify a signed challenge and authenticate the user.
        Returns JWT tokens for subsequent API calls.
      operationId: verifyChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - did
                - challenge
                - signature
                - publicKey
              properties:
                did:
                  type: string
                  description: Decentralized Identifier
                challenge:
                  type: string
                  description: Challenge from /challenge endpoint
                signature:
                  type: string
                  description: Hex-encoded signature of the message
                publicKey:
                  type: string
                  description: Hex-encoded public key for verification
            example:
              did: "did:ethr:0x1234567890abcdef1234567890abcdef12345678"
              challenge: "a1b2c3d4e5f6..."
              signature: "9f8e7d6c5b4a..."
              publicKey: "04abc123..."
      responses:
        '200':
          description: Authentication successful
          headers:
            Set-Cookie:
              description: HTTP-only authentication cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: JWT access token (7 days)
                  refreshToken:
                    type: string
                    description: JWT refresh token (30 days)
                  expiresIn:
                    type: integer
                    description: Token expiry in seconds
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      did:
                        type: string
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIs..."
                refreshToken: "eyJhbGciOiJIUzI1NiIs..."
                expiresIn: 604800
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  did: "did:ethr:0x1234567890abcdef1234567890abcdef12345678"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/auth/did/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Refresh an expired access token using a valid refresh token.
      operationId: refreshToken
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: JWT refresh token (can also be sent via cookie)
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/auth/did/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: |
        Invalidate the current session and clear authentication cookies.
      operationId: logout
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Logged out successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/auth/me:
    get:
      tags:
        - User
      summary: Get current user profile
      description: |
        Retrieve the authenticated user's profile information.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - User
      summary: Update user profile
      description: |
        Update the authenticated user's profile information.
      operationId: updateCurrentUser
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                walletAddress:
                  type: string
                walletType:
                  type: string
                  enum: [metamask, walletconnect, universal-wallet]
            example:
              email: "user@example.com"
              name: "John Doe"
              walletAddress: "0x1234567890abcdef1234567890abcdef12345678"
              walletType: "metamask"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Conflict (e.g., email already in use)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/did/verify

    CookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
      description: HTTP-only cookie set by the server

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        did:
          type: string
          description: Decentralized Identifier
        email:
          type: string
          format: email
          nullable: true
        name:
          type: string
          nullable: true
        walletAddress:
          type: string
          nullable: true
        walletType:
          type: string
          enum: [metamask, walletconnect, universal-wallet]
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
      example:
        id: "550e8400-e29b-41d4-a716-446655440000"
        did: "did:ethr:0x1234567890abcdef1234567890abcdef12345678"
        email: "user@example.com"
        name: "John Doe"
        walletAddress: "0x1234567890abcdef1234567890abcdef12345678"
        walletType: "metamask"
        createdAt: "2024-10-01T00:00:00.000Z"
        lastLoginAt: "2024-10-08T12:00:00.000Z"

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          description: Validation error details
          items:
            type: object
        retryAfter:
          type: integer
          description: Seconds to wait before retrying (for rate limit errors)

  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            details:
              - field: "did"
                message: "DID is required"

    Unauthorized:
      description: Unauthorized - authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized. Please authenticate with a valid DID."

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "User not found"

    TooManyRequests:
      description: Too many requests - rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests allowed
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Requests remaining
        X-RateLimit-Reset:
          schema:
            type: string
            format: date-time
          description: When the rate limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Too many authentication attempts. Please try again later."
            retryAfter: 60

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"